refactor app a little bit to make it cleaner
